Add files via upload
This clear the concept of multivariate regression in machine learning. Air Pollution data set is used in this project for prediction. Accuracy on test dataset got is 96% which is very good.
Code is from scratch. Recommend to go through the Andrew NG course coursera for the brief knowledge of the gradient decent and convergance.